--- a/CHANGELOG
+++ b/CHANGELOG
@@ -9,6 +9,7 @@ If not stated, FINUFFT is assumed (cuFINUFFT <=1.3 is listed separately).
 * MAX_NF increased from 1e11 to 1e12, since machines grow.
 * improved GPU python docs: migration guide; usage from cupy, numba, torch,
   pycuda. PyPI pkg still at 2.2.0beta.
+* Used new foldrescale and removed tests for the range
 
 V 2.2.0 (12/12/23)