--- a/devel/CMakeLists.txt
+++ b/devel/CMakeLists.txt
@@ -19,3 +19,4 @@ endif()
 
 add_executable(foldrescale foldrescale.cpp)
 target_link_libraries(foldrescale finufft benchmark)
+target_compile_options(foldrescale PRIVATE -mavx2)