--- a/docs/c2d.doc
+++ b/docs/c2d.doc
@@ -23,7 +23,7 @@
    Inputs:
      ntr    how many transforms (only for vectorized "many" functions, else ntr=1)
      M      number of nonuniform point sources
-     x,y    nonuniform point coordinates in [-3pi,3pi)^2 (length M real arrays)
+     x,y    nonuniform point coordinates (length M real arrays)
      c      source strengths (size M*ntr complex array)
      iflag  if >=0, uses +i in complex exponential, otherwise -i
      eps    desired relative precision; smaller is slower. This can be chosen
@@ -65,7 +65,7 @@
    Inputs:
      ntr    how many transforms (only for vectorized "many" functions, else ntr=1)
      M      number of nonuniform point targets
-     x,y    nonuniform point coordinates in [-3pi,3pi)^2 (length M real arrays)
+     x,y    nonuniform point coordinates (length M real arrays)
      iflag  if >=0, uses +i in complex exponential, otherwise -i
      eps    desired relative precision; smaller is slower. This can be chosen
             from 1e-1 down to ~ 1e-14 (in double precision) or 1e-6 (in single)