--- a/docs/cguru.doc
+++ b/docs/cguru.doc
@@ -75,7 +75,7 @@
  
    Notes:
      * For type 1 and 2, the values in x (and if nonempty, y and z) can be in any
-       interval, they will be folded to [-pi, pi]. Note: for large numbers outside
+       interval, they will be folded to [-pi, pi). Note: for large numbers outside
        this range there might be numerical inaccuracies.
        For type 1 they are "sources", but for type 2, "targets".
        In contrast, for type 3 there are no restrictions on them,