… CMakeLists.txt
… CMakePresets.json
… docs/c_gpu.rst
… docs/cex.html
… docs/error.rst
… docs/index.html
… docs/matlabhelp.doc
… docs/opts.rst
… examples/cuda/getting_started.cpp
… fortran/examples/guru1d1.f
… fortran/examples/guru1d1f.f
… fortran/examples/simple1d1.f
… fortran/examples/simple1d1.f90
… fortran/examples/simple1d1f.f
… include/finufft_errors.h
… matlab/finufft2d1.m
… matlab/finufft2d2.m
… matlab/finufft3d1.m
… matlab/finufft3d2.m
… matlab/finufft_plan.m
… matlab/opts12.docbit
… perftest/spreadtestnd.cpp
… python/cufinufft/cufinufft/_simple.py
… python/finufft/finufft/_interfaces.py
… python/finufft/test/test_finufft_plan.py
… src/cuda/1d/interp1d_wrapper.cu
… src/cuda/1d/spread1d_wrapper.cu
… src/cuda/1d/spreadinterp1d.cuh
… src/cuda/2d/interp2d_wrapper.cu
… src/cuda/2d/spread2d_wrapper.cu