Optimized foldrescale using @mreineck suggestions.  This removed the range limitation and made the code faster. Results in the comments. 

List of changes:

- [x] Integrate @mreineck version
- [x] Change from macro to inline function
- [x] remove bounds checking
- [x] deprecate chkbnds 
- [x] update docs for chkbnds
- [x] change tests to remove chkbnds
- [x] remove pirange
- [x] update docs for pirange