@DiamonDinoia 2024-05-08
Performance: `./spreadtestnd 3 1e07 10e7 1e-6 1 0 1`
Before:
```
setup_spreader (kerevalmeth=1) eps=1e-06 sigma=2: chose ns=7 beta=16.1
	sorted (1 threads):	0.00121 s
	spread 3D (M=1; N1=464,N2=464,N3=464; pir=0), nthr=1
	zero output array	0.212 s
	using low-density speed rescue nb=M...
	t1 fancy spread: 	3e-05 s (1 subprobs)
making random data...
spreadinterp 3D, 9.99e+07 U pts, dir=1, tol=1e-06: nspread=7
	sorted (1 threads):	0.332 s
	spread 3D (M=10000000; N1=464,N2=464,N3=464; pir=0), nthr=1
	zero output array	0.212 s
	capping subproblem sizes to max of 100000
	t1 fancy spread: 	6.45 s (100 subprobs)
    1e+07 NU pts in 7 s 	1.43e+06 pts/s 	4.9e+08 spread pts/s
    rel err in total over grid:      3.28e-07
making more random NU pts...
spreadinterp 3D, 9.99e+07 U pts, dir=2, tol=1e-06: nspread=7
	sorted (1 threads):	0.364 s
	interp 3D (M=10000000; N1=464,N2=464,N3=464; pir=0), nthr=1
	t2 spreading loop: 	4.75 s
    1e+07 NU pts in 5.11 s 	1.96e+06 pts/s 	6.71e+08 spread pts/s
    max rel err in values at NU pts: 3.39e-06
```
Using the new FOLDRESCALE, force inlining:
```
setup_spreader (kerevalmeth=1) eps=1e-06 sigma=2: chose ns=7 beta=16.1
	sorted (1 threads):	0.00127 s
	spread 3D (M=1; N1=464,N2=464,N3=464; pir=0), nthr=1
	zero output array	0.216 s
	using low-density speed rescue nb=M...
	t1 fancy spread: 	2.9e-05 s (1 subprobs)
making random data...
spreadinterp 3D, 9.99e+07 U pts, dir=1, tol=1e-06: nspread=7
	sorted (1 threads):	0.314 s
	spread 3D (M=10000000; N1=464,N2=464,N3=464; pir=0), nthr=1
	zero output array	0.215 s
	capping subproblem sizes to max of 100000
	t1 fancy spread: 	5.8 s (100 subprobs)
    1e+07 NU pts in 6.33 s 	1.58e+06 pts/s 	5.42e+08 spread pts/s
    rel err in total over grid:      3.28e-07
making more random NU pts...
spreadinterp 3D, 9.99e+07 U pts, dir=2, tol=1e-06: nspread=7
	sorted (1 threads):	0.313 s
	interp 3D (M=10000000; N1=464,N2=464,N3=464; pir=0), nthr=1
	t2 spreading loop: 	4.75 s
    1e+07 NU pts in 5.07 s 	1.97e+06 pts/s 	6.77e+08 spread pts/s
    max rel err in values at NU pts: 3.39e-06
```